kamping/requirements.txt
amqp==2.5.2
billiard==3.6.1.0
colorama==0.4.1
Django==2.2.8
django-celery-beat==1.5.0
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-timezone-field==3.1
django-vote==2.1.7
feedback==0.1.post7
importlib-metadata==0.22
kombu
more-itertools==7.2.0
Pillow==6.2.1
psycopg2==2.8.3
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
Unidecode==1.1.1
vine==1.3.0
zipp==0.6.0
celery
redis
djangorestframework